only delete old dirs if present

+++ launch_backend.py
@@ -173,9 +173,12 @@ while True:
             dest = SAVING_DIR_ROOT + config_file_name[:-5]
             if os.path.exists(dest) and not should_resume:
                 print('deleting existing model dir: ', dest)
-                shutil.rmtree(dest + '/model')
-                shutil.rmtree(dest + '/tensorboard')
-                shutil.rmtree(dest + '/other_logs')
+                if os.path.exists(dest + '/model'):
+                    shutil.rmtree(dest + '/model')
+                if os.path.exists(dest + '/tensorboard'):
+                    shutil.rmtree(dest + '/tensorboard')
+                if os.path.exists(dest + '/other_logs'):
+                    shutil.rmtree(dest + '/other_logs')
             elif should_resume and not os.path.exists(dest):
                 warnings.warn('trying to resume training but model dir does not exist')
             saving_dir = create_saving_dir(SAVING_DIR_ROOT, config_file_name[:-5])
